+++ identity-server/app.js
@@ -2,8 +2,8 @@ const jwt = require("jsonwebtoken");
 const express = require("express");
 const bodyParser = require('body-parser');
 require("dotenv").config({ path: '../.env' });
-const dbHandler = require("./dbHandler");
-const kitchen = require("./kitchen");
+const dbHandler = require("./database/dbHandler");
+const kitchen = require("./kitchen/kitchen");
 const verify = require("./middleware/verify");
 
 const app = express();
