+++ identity-server/database/dbService.js
@@ -1,5 +1,5 @@
 const sql = require("mssql");
-require("dotenv").config({ path: '../.env' });
+require("dotenv").config({ path: '../../.env' });
 
 const connect = async () => {
